structure: 구조 분할
